snapshot circa 2009
this package used to live in the "lsware" CVS project hosted on gravity.phys.uwm.edu.  that server was taken down years ago.  this snapshot has been recovered from a working copy I had in my home directory.

I started this project in about 2007 as a proof-of-principle to demonstrate how straight-forward it would be to separate lalsuite from libmetaio, in the hope that people would see the benefit.  the idea didn't get enough up-take:  I wasn't prepared to fix other people's codes for them, they were going to have to want to improve their tools for their own reasons, and people didn't.  years later when the gstlal project needed to move forward with XML format changes, specifically changing the "ilwd:char" row IDs to integers, and the rest of the collaboration wouldn't agree to it, this code was resurrected and added to the gstlal project allowing gstlal to decouple from lalsuite and make XML table format changes independently of the collaboration.

more recently, a studnent of mine has found it necessary to read PSD arrays in C code, and, of course, that's impossible because why would anybody provide an I/O library for a file format they invented, I mean, sheesh, what do you think this is?  we could write a python translator to convert the PSDs from LIGO LW XML to something that can be read in C, like HDF5, but then an HDF5 PSD extractor would still need to be written, plus using any resulting tool would require an external translation step to be run on the PSDs first which is inconvenient.  it's probably the same total lines of code to just add the capability directly in C here, and the result will be more convenient to use.  so that's what I'm doing.